requirement.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.6.0
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.15
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-extensions==3.2.1
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.13.1
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
importlib-metadata==4.13.0
ipdb==0.13.11
ipython==7.34.0
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
protobuf==4.21.12
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
redis==4.3.4
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
zipp==3.12.0
